modified:   UNITE_Main.X/SampleManager.c 	modified:   UNITE_Main.X/SatelliteMode.c 	modified:   UNITE_Main.X/SatelliteMode.h 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/SatelliteMode.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/system.o 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.elf 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.hex 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/UNITE_Main.X.production.map 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/memoryfile.xml 	modified:   UNITE_Main.X/nbproject/private/private.xml 	modified:   UNITE_Main.X/system.c 	modified:   UNITE_Main.X/system.h

+++ UNITE_Main.X/system.c
@@ -101,8 +101,8 @@ void wait_for(unsigned long something) {
     for (i = 0; i < wait; i++);
 }
 
-uint16_t Pow(int base, int exponent) {
-  uint16_t returnValue = 1;
+unsigned long Pow(int base, int exponent) {
+  unsigned long returnValue = 1;
   int i;
   for (i = 0; i < exponent; i++) {
       returnValue = returnValue * base;
